Moved: Writer to world from writer.py
