+++ tests/objects/test_writer.py
@@ -8,14 +8,14 @@ import unittest
 from tests.testutils import print_testtitle, validate_with_fail
 from builder.commands.scode import SCode, SCmd
 from builder.objects.sobject import SObject
-from builder.objects import writer as wt
+from builder.world import Writer
 
 
 class WriterTest(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        print_testtitle(wt.__name__, 'Writer class')
+        print_testtitle("world.Writer", 'Writer class')
 
     def test_instance(self):
         taro = SObject('taro')
@@ -45,8 +45,8 @@ class WriterTest(unittest.TestCase):
                     taro, SCmd.WEAR, ('a',)),
                 ]
         def checker(src, act, args, exp_src, exp_cmd, exp_args):
-            tmp = wt.Writer(src)
-            self.assertIsInstance(tmp, wt.Writer)
+            tmp = Writer(src)
+            self.assertIsInstance(tmp, Writer)
             tmp2 = getattr(tmp, act)(*args)
             self.assertIsInstance(tmp2, SCode)
             self.assertEqual(tmp2.src, exp_src)
